Add files via upload
This has been an updated version of my current layout, When adding logic i reduced my current pizza size to one, but i have updated the chamber of the 5 cheeses to have a border around their image, I have also added a submit button that works to give me a name and email that prints on the IDE when we run the server. I still need to work on the feedback portion of my interaction, but with the enhanced CSS, HTML, and updated pizza console logic, and the JavaScript server, this is a good reset on the foundation, where now I can add more pizzas in a more dynamic way rather than adding them one after another. This is just the start to what is a good first server and info, but I want to add much more and will continue to add more throughout.